+++ airwaves_adsb_client.py
@@ -10,7 +10,6 @@ import time
 import datetime
 from libPyAirwaves.adsb import is_valid_adsb_message
 from libPyAirwaves.structs import AdsbType
-from models import Session, Aircrafts, AircraftModes, AircraftOwner, AircraftRegistration
 import redis
 import json
 
@@ -20,7 +19,6 @@ max_failed_connection = 10
 redis = redis.from_url(config.REDIS_URL)
 pubsub = redis.pubsub()
 pubsub.subscribe("room:vehicles")
-session = Session()
 
 if __name__ == "__main__":
     while count_failed_connection < max_failed_connection:
@@ -88,49 +86,20 @@ if __name__ == "__main__":
                 if len(line) == 22:
                     if is_valid_adsb_message(line):
                         adsb_message = AdsbType()
-                        adsb_message.populate_from_list(line)
+                        adsb_message.populate_from_sbs(line)
+                        # print(adsb_message.to_dict())
                         # Emit Socket.IO message only if altitude, latitude and longitude are set
                         # AKA a "MSG,3" message and perhaps a "MSG,2" (Surface position)
                         if adsb_message.has_location():
                             adsb_message.entryPoint = "airwaves_adsb_client"
-                            adsb_message.src = config.PYAW_HOSTNAME
-                            adsb_message.clientName = config.ADSB_SOURCE["name"]
+                            adsb_message.ourName = config.PYAW_HOSTNAME
+                            adsb_message.srcName = config.ADSB_SOURCE["name"]
+                            adsb_message.srcLat = config.ADSB_SOURCE["lat"]
+                            adsb_message.srcLon = config.ADSB_SOURCE["lon"]
+                            adsb_message.srcPosMode = config.ADSB_SOURCE["posMode"]
                             adsb_message.dataOrigin = "dump1090"
 
-                            # Get more datas from SQL
-                            sqlcraft = (
-                                session.query(
-                                    AircraftModes.icao_type_code,
-                                    Aircrafts.type,
-                                    Aircrafts.manufacturer,
-                                    Aircrafts.aircraft_description,
-                                    Aircrafts.engine_type,
-                                    Aircrafts.engine_count,
-                                    AircraftOwner.registration,
-                                    AircraftOwner.owner,
-                                    AircraftRegistration.country,
-                                    AircraftRegistration.prefix,
-                                )
-                                .filter(AircraftModes.mode_s == adsb_message.addr)
-                                .join(Aircrafts, Aircrafts.icao == AircraftModes.icao_type_code)
-                                .join(AircraftOwner, AircraftOwner.registration == AircraftModes.registration)
-                                .join(
-                                    AircraftRegistration, AircraftRegistration.country == AircraftModes.mode_s_country
-                                )
-                                .first()
-                            )
-
-                            try:
-                                adsb_message.icaoAACC = sqlcraft.country
-                            except AttributeError:
-                                None
-
-                            try:
-                                adsb_message.category = sqlcraft.aircraft_description
-                            except AttributeError:
-                                None
-
-                            # print(adsb_message.to_dict())
+                            print(adsb_message.to_dict())
                             redis.publish("room:vehicles", json.dumps(adsb_message.to_dict()))
                     # It's valid, reset stream message
                     data_str = ""