+++ backend/lib/pyairwaves_web/endpoint.ex
@@ -1,5 +1,6 @@
 defmodule PyairwavesWeb.Endpoint do
   use Phoenix.Endpoint, otp_app: :pyairwaves
+  use Sentry.Phoenix.Endpoint
 
   # The session will be stored in the cookie and signed,
   # this means its contents can be read but not tampered with.