+++ backend/lib/pyairwaves_web/views/archive_view.ex
@@ -1,0 +1,3 @@
+defmodule PyairwavesWeb.ArchiveView do
+  use PyairwavesWeb, :view
+end