+++ backend/priv/static/js/core/messages.js
@@ -16,7 +16,8 @@
  **************************************************/
 
 function handleMessage(event) {
-    const msgJSON = JSON.parse(event.data);
+    // const msgJSON = JSON.parse(event.data);
+    const msgJSON = event.data;
     if (!mapLoaded) {
         // Return early - the map isn't loaded yet
         Logger.error("Maps not loaded. Discarding aircraft data.");