+++ simulator_adsb_coordinates.py
@@ -31,7 +31,11 @@ try:
         adsb_msg = AdsbType()
         adsb_msg.populate_from_list(msg)
         adsb_msg.entryPoint = "simulator"
-        adsb_msg.src = cfg.PYAW_HOSTNAME
+        adsb_msg.ourName = cfg.PYAW_HOSTNAME
+        adsb_msg.srcName = cfg.ADSB_SOURCE["name"]
+        adsb_msg.srcLat = cfg.ADSB_SOURCE["lat"]
+        adsb_msg.srcLon = cfg.ADSB_SOURCE["lon"]
+        adsb_msg.srcPosMode = cfg.ADSB_SOURCE["posMode"]
         adsb_msg.clientName = "sim_host"
         adsb_msg.dataOrigin = "dump1090"
         redis.publish("room:vehicles", json.dumps(adsb_msg.to_dict()))
@@ -48,7 +52,11 @@ try:
         adsb_msg = AdsbType()
         adsb_msg.populate_from_list(msg)
         adsb_msg.entryPoint = "simulator"
-        adsb_msg.src = cfg.PYAW_HOSTNAME
+        adsb_msg.ourName = cfg.PYAW_HOSTNAME
+        adsb_msg.srcName = cfg.ADSB_SOURCE["name"]
+        adsb_msg.srcLat = cfg.ADSB_SOURCE["lat"]
+        adsb_msg.srcLon = cfg.ADSB_SOURCE["lon"]
+        adsb_msg.srcPosMode = cfg.ADSB_SOURCE["posMode"]
         adsb_msg.clientName = "sim_host"
         adsb_msg.dataOrigin = "dump1090"
         redis.publish("room:vehicles", json.dumps(adsb_msg.to_dict()))
@@ -65,7 +73,11 @@ try:
         adsb_msg = AdsbType()
         adsb_msg.populate_from_list(msg)
         adsb_msg.entryPoint = "simulator"
-        adsb_msg.src = cfg.PYAW_HOSTNAME
+        adsb_msg.ourName = cfg.PYAW_HOSTNAME
+        adsb_msg.srcName = cfg.ADSB_SOURCE["name"]
+        adsb_msg.srcLat = cfg.ADSB_SOURCE["lat"]
+        adsb_msg.srcLon = cfg.ADSB_SOURCE["lon"]
+        adsb_msg.srcPosMode = cfg.ADSB_SOURCE["posMode"]
         adsb_msg.clientName = "sim_host"
         adsb_msg.dataOrigin = "dump1090"
         redis.publish("room:vehicles", json.dumps(adsb_msg.to_dict()))