+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- AIS Messages logging
+
 ### Changed
 - **Breaking:** Backend rewritten from python to Elixir
 - Frontend: Markers out of the current displayed map bounds are dropped when received
+- Renamed srcPosMeta to srcPosMode
 
 ## [0.0.1]
 ### Added